Split sync dependency configuration
