+++ tests/test-openrouter.js
@@ -286,6 +286,7 @@ assert('startup-oauth-callbacks.js clears pending OAuth state after callback',
 assert('startup-oauth-callbacks.js marks fresh OpenRouter settings local for sync',
   startupOAuthSrc.includes('markOpenRouterOAuthSettingsLocal()'));
 const syncSrc = read('js/sync.js');
+const syncConfigureSrc = read('js/sync-configure.js');
 const syncReconcileSrc = read('js/sync-reconcile.js');
 const syncApplySrc = read('js/sync-apply.js');
 assert('sync preserves fresh OpenRouter OAuth provider/key against stale pull',
@@ -295,7 +296,8 @@ assert('sync preserves fresh local AI settings against stale pull',
 assert('sync refreshes AI header after remote AI settings apply',
   syncApplySrc.includes('window.updateChatHeaderModel?.()') && syncApplySrc.includes('window.refreshWebSearchToggle?.()'));
 assert('startup sync reconciliation pushes local AI setting drift',
-  syncSrc.includes("from './sync-reconcile.js'")
+  syncConfigureSrc.includes("from './sync-reconcile.js'")
+    && syncSrc.includes("from './sync-configure.js'")
     && syncReconcileSrc.includes('newer local AI settings')
     && syncReconcileSrc.includes('collectAISettings()'));
 const cssSrc = read('styles.css');
